[selected-names-list] selectedNames refactored and fully functional
